engine: move pb receiver in own file
No code changes.
