+++ storage/engine/mvcc.go
@@ -122,135 +122,6 @@ type MVCCKeyValue struct {
 	Value []byte
 }
 
-// Value returns the inline value.
-func (meta MVCCMetadata) Value() roachpb.Value {
-	return roachpb.Value{RawBytes: meta.RawBytes}
-}
-
-// IsInline returns true if the value is inlined in the metadata.
-func (meta MVCCMetadata) IsInline() bool {
-	return meta.RawBytes != nil
-}
-
-// IsIntentOf returns true if the meta record is an intent of the supplied
-// transaction.
-func (meta MVCCMetadata) IsIntentOf(txn *roachpb.Transaction) bool {
-	return meta.Txn != nil && txn != nil && roachpb.TxnIDEqual(meta.Txn.ID, txn.ID)
-}
-
-// GCBytes is a convenience function which returns the number of gc bytes,
-// that is the key and value bytes excluding the live bytes.
-func (ms MVCCStats) GCBytes() int64 {
-	return ms.KeyBytes + ms.ValBytes - ms.LiveBytes
-}
-
-// Total returns the range size as the sum of the key and value
-// bytes. This includes all non-live keys and all versioned values.
-func (ms MVCCStats) Total() int64 {
-	return ms.KeyBytes + ms.ValBytes
-}
-
-// AgeTo encapsulates the complexity of computing the increment in age
-// quantities contained in MVCCStats. Two MVCCStats structs only add and
-// subtract meaningfully if their LastUpdateNanos matches, so aging them to
-// the max of their LastUpdateNanos is a prerequisite.
-// If nowNanos is behind ms.LastUpdateNanos, this method is a noop.
-func (ms *MVCCStats) AgeTo(nowNanos int64) {
-	if ms.LastUpdateNanos >= nowNanos {
-		return
-	}
-	diffSeconds := nowNanos/1E9 - ms.LastUpdateNanos/1E9 // not (...)/1E9!
-
-	ms.GCBytesAge += ms.GCBytes() * diffSeconds
-	ms.IntentAge += ms.IntentCount * diffSeconds
-	ms.LastUpdateNanos = nowNanos
-}
-
-// Add adds values from oms to ms. The ages will be moved forward to the
-// larger of the LastUpdateNano timestamps involved.
-func (ms *MVCCStats) Add(oms MVCCStats) {
-	// Enforce the max LastUpdateNanos for both ages based on their
-	// pre-addition state.
-	ms.AgeTo(oms.LastUpdateNanos)
-	oms.AgeTo(ms.LastUpdateNanos)
-	// Now that we've done that, we may just add them.
-	ms.IntentAge += oms.IntentAge
-	ms.GCBytesAge += oms.GCBytesAge
-
-	ms.LiveBytes += oms.LiveBytes
-	ms.KeyBytes += oms.KeyBytes
-	ms.ValBytes += oms.ValBytes
-	ms.IntentBytes += oms.IntentBytes
-	ms.LiveCount += oms.LiveCount
-	ms.KeyCount += oms.KeyCount
-	ms.ValCount += oms.ValCount
-	ms.IntentCount += oms.IntentCount
-	ms.SysBytes += oms.SysBytes
-	ms.SysCount += oms.SysCount
-}
-
-// Subtract removes oms from ms. The ages will be moved forward to the larger of
-// the LastUpdateNano timestamps involved.
-func (ms *MVCCStats) Subtract(oms MVCCStats) {
-	// Enforce the max LastUpdateNanos for both ages based on their
-	// pre-subtraction state.
-	ms.AgeTo(oms.LastUpdateNanos)
-	oms.AgeTo(ms.LastUpdateNanos)
-	// Now that we've done that, we may subtract.
-	ms.IntentAge -= oms.IntentAge
-	ms.GCBytesAge -= oms.GCBytesAge
-	ms.LiveBytes -= oms.LiveBytes
-	ms.KeyBytes -= oms.KeyBytes
-	ms.ValBytes -= oms.ValBytes
-	ms.IntentBytes -= oms.IntentBytes
-	ms.LiveCount -= oms.LiveCount
-	ms.KeyCount -= oms.KeyCount
-	ms.ValCount -= oms.ValCount
-	ms.IntentCount -= oms.IntentCount
-	ms.SysBytes -= oms.SysBytes
-	ms.SysCount -= oms.SysCount
-}
-
-// AvgIntentAge returns the average age of outstanding intents,
-// based on current wall time specified via nowNanos.
-func (ms MVCCStats) AvgIntentAge(nowNanos int64) float64 {
-	if ms.IntentCount == 0 {
-		return 0
-	}
-	// Advance age by any elapsed time since last computed. Note that
-	// we operate on a copy.
-	ms.AgeTo(nowNanos)
-	return float64(ms.IntentAge) / float64(ms.IntentCount)
-}
-
-// GCByteAge returns the total age of outstanding gc'able
-// bytes, based on current wall time specified via nowNanos.
-// nowNanos is ignored if it's a past timestamp as seen by
-// rs.LastUpdateNanos.
-func (ms MVCCStats) GCByteAge(nowNanos int64) int64 {
-	ms.AgeTo(nowNanos) // we operate on a copy
-	return ms.GCBytesAge
-}
-
-// AccountForSelf adjusts ms to account for the predicted impact it will have on
-// the values that it records when the structure is initially stored. Specifically,
-// MVCCStats is stored on the RangeStats key, which means that its creation will
-// have an impact on system-local data size and key count.
-func (ms *MVCCStats) AccountForSelf(rangeID roachpb.RangeID) error {
-	key := keys.RangeStatsKey(rangeID)
-	metaKey := MakeMVCCMetadataKey(key)
-
-	// MVCCStats is stored inline, so compute MVCCMetadata accordingly.
-	value := roachpb.Value{}
-	if err := value.SetProto(ms); err != nil {
-		return err
-	}
-	meta := MVCCMetadata{RawBytes: value.RawBytes}
-
-	updateStatsForInline(ms, key, 0, 0, int64(metaKey.EncodedSize()), int64(meta.Size()))
-	return nil
-}
-
 // isSysLocal returns whether the whether the key is system-local.
 func isSysLocal(key roachpb.Key) bool {
 	return key.Compare(keys.LocalMax) < 0
